+++ CHANGELOG.md
@@ -1,6 +1,12 @@
 Changelog
 =========
 
+* **2013-08-08**:
+  * Seperate Multilang document now incorporated in single document
+  * `multilang.locales` config removed (use cmf_core.multilang.locales)
+  * `DataFixtures/LoadCmsData` renamed and moved to DataFixtures/Phpcr
+  * PHPCR documents moved from Document to Doctrine/Phpcr
+
 * **2013-08-04**: Changed name of Sonata route names / patterns - now /admin/cmf/simplecms/foo instead of /admin/bundle/simplecms/foo
 
 1.0.0-beta3